requirements.txt
Flask==3.0.3
requests==2.32.3
beautifulsoup4==4.12.3
python-dateutil==2.9.0
bootstrap-flask==2.4.0
mysql-connector-python==8.4.0
python-dotenv==1.0.1